[task] Added graphing of traceroute with analysis
Added module to read csv file and graph trace the network topology.
It compares two given times and draws the graphs with the analysis
for each. It includes average rtt between hops, the standard deviation
and variance to tell how stable the route node in the route is, and lastly
the number of times the node failed to respond.